refactor(server-core): move the openapi coverage guards into the test suite
The enrichment step had grown into two unrelated jobs: producing part of the
document, and asserting things about the code that produced it. The second is
an invariant, not generation, and this repository keeps invariants in specs.

So the marker cross-checks moved to `test/unit/http/openapi-coverage.spec.ts`,
next to the parity spec that already makes this class of claim. They are better
off there. The source half needs no built document, so it runs on every suite
run rather than only when someone regenerates; the document half skips cleanly
when `dist/swagger.json` is absent, which `build:server:js` makes an ordinary
state. A failure now reports as a failing assertion with a diff instead of a
thrown build error, and the checks can be exercised directly, which is how the
port was verified: mismarking a route fails the spec with the same message the
build used to print.

What is left in the script is only what produces the document, and it no longer
reads a file or knows a registry: it is a pure function over what trapi emitted,
which is what let the `node:fs`, `node:path` and `node:url` imports go with the
guards. 770 lines when it was a chained build step, 437 now.

It exists at all because two things are not expressible in trapi yet: a
decorator handler cannot attach the parameters it derives (tada5hi/trapi#907,
the half of #895 that #899 left), and there is no way to declare a response
that applies to every operation (tada5hi/trapi#908). With both, the parameters
move into the handler that already resolves the schema, the error response
becomes config, and this file goes away entirely.
